minor fixes

+++ src/main/java/com/nuach_server/llm/OllamaOutput.java
@@ -1,17 +1,19 @@
 package com.nuach_server.llm;
 
+import java.math.BigInteger;
+
 public record OllamaOutput(
         String model,
         String created_at,
         String response,
         Boolean done,
         String done_reason,
-        int[]  context,
-        int total_duration,
-        int load_duration,
-        int prompt_eval_count,
-        int prompt_eval_duration,
-        int eval_count,
-        int eval_duration
+        BigInteger[]  context,
+        BigInteger total_duration,
+        BigInteger load_duration,
+        BigInteger prompt_eval_count,
+        BigInteger prompt_eval_duration,
+        BigInteger eval_count,
+        BigInteger eval_duration
 ) {
 }
